fix(typescript): correctly render the types of nodes
This fix also includes a refactoring that allows the developer to
override the default behaviour of this rendering by providing their
own `Host` implementation.

BREAKING CHANGE:

The way that some types are rendered has changed. It is believed that
this is better but you may want to check in case your expectation is
different.

* `let myVar = 42` used to render the type as `42` but
will now render as `number`.

* type parameters of the following call signature:

```
<T, U extends Findable<T>>(param: T) : U
```

will be rendered as `<T, U>` whereas before it might be have been
rendered as `<T, U extends Findable<T>>`.

* structural types are now rendered as a single line:

```
{ foo: string; }
```

rather than

```
{
    foo: string;
}
```

* Results of complex calls might become `any` whereas, previously,
the call itself was rendered. For example:

```
Object.assign(super.someProp, {
  bar: 'baz'
});
```

Fixes #276